+++ src/Api/Endpoints/Documenten/Enkelvoudiginformatieobjecten.php
@@ -13,9 +13,17 @@ use Woweb\Zgw\Api\Actions\Search;
 use Woweb\Zgw\Api\Actions\Show;
 use Woweb\Zgw\Api\Actions\Store;
 use Woweb\Zgw\Api\Endpoints\AbstractEndpoint;
+use Woweb\Zgw\Contracts\CreatesResource;
+use Woweb\Zgw\Contracts\DeletesResource;
+use Woweb\Zgw\Contracts\ListsResources;
+use Woweb\Zgw\Contracts\PatchesResource;
+use Woweb\Zgw\Contracts\ProvidesAuditTrail;
+use Woweb\Zgw\Contracts\ReplacesResource;
+use Woweb\Zgw\Contracts\SearchesResources;
+use Woweb\Zgw\Contracts\ShowsResource;
 use Woweb\Zgw\Exceptions\ApiRequestException;
 
-class Enkelvoudiginformatieobjecten extends AbstractEndpoint
+class Enkelvoudiginformatieobjecten extends AbstractEndpoint implements CreatesResource, DeletesResource, ListsResources, PatchesResource, ProvidesAuditTrail, ReplacesResource, SearchesResources, ShowsResource
 {
     use Audittrail;
     use Delete;