+++ src/Api/Endpoints/Zaken/Klantcontacten.php
@@ -8,8 +8,11 @@ use Woweb\Zgw\Api\Actions\Index;
 use Woweb\Zgw\Api\Actions\Show;
 use Woweb\Zgw\Api\Actions\Store;
 use Woweb\Zgw\Api\Endpoints\AbstractEndpoint;
+use Woweb\Zgw\Contracts\CreatesResource;
+use Woweb\Zgw\Contracts\ListsResources;
+use Woweb\Zgw\Contracts\ShowsResource;
 
-class Klantcontacten extends AbstractEndpoint
+class Klantcontacten extends AbstractEndpoint implements CreatesResource, ListsResources, ShowsResource
 {
     use Index;
     use Show;