+++ src/Api/Endpoints/Zaken/Zaken.php
@@ -15,8 +15,16 @@ use Woweb\Zgw\Api\Actions\Store;
 use Woweb\Zgw\Api\Endpoints\AbstractEndpoint;
 use Woweb\Zgw\Api\Endpoints\Zaken\Nested\ZaakBesluiten;
 use Woweb\Zgw\Api\Endpoints\Zaken\Nested\Zaakeigenschappen;
+use Woweb\Zgw\Contracts\CreatesResource;
+use Woweb\Zgw\Contracts\DeletesResource;
+use Woweb\Zgw\Contracts\ListsResources;
+use Woweb\Zgw\Contracts\PatchesResource;
+use Woweb\Zgw\Contracts\ProvidesAuditTrail;
+use Woweb\Zgw\Contracts\ReplacesResource;
+use Woweb\Zgw\Contracts\SearchesResources;
+use Woweb\Zgw\Contracts\ShowsResource;
 
-class Zaken extends AbstractEndpoint
+class Zaken extends AbstractEndpoint implements CreatesResource, DeletesResource, ListsResources, PatchesResource, ProvidesAuditTrail, ReplacesResource, SearchesResources, ShowsResource
 {
     use Audittrail;
     use Delete;